Fix: move NAT instance to tunnel
